requirements.txt
discord.py==2.1.0
pillow==9.3.0
numpy==1.23.4
python-dateutil==2.8.2
PyYAML==6.0
aiomysql==0.1.1
PyMySQL==1.0.2
psutil==5.9.4
pytz==2022.6
pandas==1.5.1
wand==0.6.10
aiohttp==3.8.3
